Update Wabt for shared memory support Use wabt validator when converting wast to wasm
