+++ lib/Runtime/Library/WabtInterface.cpp
@@ -144,7 +144,7 @@ Js::Var WabtInterface::EntryConvertWast2Wasm(RecyclableObject* function, CallInf
     try
     {
         ChakraWabt::SpecContext spec;
-        ChakraWabt::Context wabtCtx;
+        ChakraWabt::ChakraContext wabtCtx;
         wabtCtx.user_data = &context;
         wabtCtx.createBuffer = CreateBuffer;
         wabtCtx.features.sign_extends = CONFIG_FLAG(WasmSignExtends);
