+++ lib/wabt/src/leb128.h
@@ -56,6 +56,11 @@ void WriteS32Leb128(Stream* stream, T value, const char* desc) {
   WriteS32Leb128(stream, static_cast<uint32_t>(value), desc);
 }
 
+// Returns the length of the leb128.
+size_t ReadU32Leb128(const uint8_t* p, const uint8_t* end, uint32_t* out_value);
+size_t ReadS32Leb128(const uint8_t* p, const uint8_t* end, uint32_t* out_value);
+size_t ReadS64Leb128(const uint8_t* p, const uint8_t* end, uint64_t* out_value);
+
 }  // namespace wabt
 
 #endif // WABT_LEB128_H_
